feat(i18n): refactor the structure of i18n and complete some chinese translation
