+++ packages/drawnix/src/i18n/translations/en.ts
@@ -1,0 +1,140 @@
+import { Translations } from '../types';
+
+const enTranslations: Translations = {
+  // Toolbar items
+  'toolbar.hand': 'Hand — H',
+  'toolbar.selection': 'Selection — V',
+  'toolbar.mind': 'Mind — M',
+  'toolbar.text': 'Text — T',
+  'toolbar.arrow': 'Arrow — A',
+  'toolbar.shape': 'Shape',
+  'toolbar.image': 'Image — Cmd+U',
+  'toolbar.extraTools': 'Extra Tools',
+
+  'toolbar.pen': 'Pen — P',
+  'toolbar.eraser': 'Eraser — E',
+
+  'toolbar.arrow.straight': 'Straight Arrow Line',
+  'toolbar.arrow.elbow': 'Elbow Arrow Line',
+  'toolbar.arrow.curve': 'Curve Arrow Line',
+
+  'toolbar.shape.rectangle': 'Rectangle — R',
+  'toolbar.shape.ellipse': 'Ellipse — O',
+  'toolbar.shape.triangle': 'Triangle',
+  'toolbar.shape.terminal': 'Terminal',
+  'toolbar.shape.diamond': 'Diamond',
+  'toolbar.shape.parallelogram': 'Parallelogram',
+  'toolbar.shape.roundRectangle': 'Round Rectangle',
+
+  // Zoom controls
+  'zoom.in': 'Zoom In — Cmd++',
+  'zoom.out': 'Zoom Out — Cmd+-',
+  'zoom.fit': 'Fit to Screen',
+  'zoom.100': 'Zoom to 100%',
+
+  // Themes
+  'theme.default': 'Default',
+  'theme.colorful': 'Colorful',
+  'theme.soft': 'Soft',
+  'theme.retro': 'Retro',
+  'theme.dark': 'Dark',
+  'theme.starry': 'Starry',
+
+  // Colors
+  'color.none': 'Topic Color',
+  'color.unknown': 'Other Color',
+  'color.default': 'Basic Black',
+  'color.white': 'White',
+  'color.gray': 'Grey',
+  'color.deepBlue': 'Deep Blue',
+  'color.red': 'Red',
+  'color.green': 'Green',
+  'color.yellow': 'Yellow',
+  'color.purple': 'Purple',
+  'color.orange': 'Orange',
+  'color.pastelPink': 'Paster Pink',
+  'color.cyan': 'Cyan',
+  'color.brown': 'Brown',
+  'color.forestGreen': 'Forest Green',
+  'color.lightGray': 'Light Grey',
+
+  // General
+  'general.undo': 'Undo',
+  'general.redo': 'Redo',
+  'general.menu': 'App Menu',
+  'general.duplicate': 'Duplicate',
+  'general.delete': 'Delete',
+
+  // Language
+  'language.switcher': 'Language',
+  'language.chinese': '中文',
+  'language.english': 'English',
+  'language.russian': 'Русский',
+
+  // Menu items
+  'menu.open': 'Open',
+  'menu.saveFile': 'Save File',
+  'menu.exportImage': 'Export Image',
+  'menu.exportImage.png': 'PNG',
+  'menu.exportImage.jpg': 'JPG',
+  'menu.cleanBoard': 'Clear Board',
+  'menu.github': 'GitHub',
+
+  // Dialog translations
+  'dialog.mermaid.title': 'Mermaid to Drawnix',
+  'dialog.mermaid.description': 'Currently supports',
+  'dialog.mermaid.flowchart': 'flowcharts',
+  'dialog.mermaid.sequence': 'sequence diagrams',
+  'dialog.mermaid.class': 'class diagrams',
+  'dialog.mermaid.otherTypes':
+    ', and other diagram types (rendered as images).',
+  'dialog.mermaid.syntax': 'Mermaid Syntax',
+  'dialog.mermaid.placeholder': 'Write your Mermaid chart definition here…',
+  'dialog.mermaid.preview': 'Preview',
+  'dialog.mermaid.insert': 'Insert',
+  'dialog.markdown.description':
+    'Supports automatic conversion of Markdown syntax to mind map.',
+  'dialog.markdown.syntax': 'Markdown Syntax',
+  'dialog.markdown.placeholder':
+    'Write your Markdown text definition here...',
+  'dialog.markdown.preview': 'Preview',
+  'dialog.markdown.insert': 'Insert',
+  'dialog.error.loadMermaid': 'Failed to load Mermaid library',
+
+  // Extra tools menu items
+  'extraTools.mermaidToDrawnix': 'Mermaid to Drawnix',
+  'extraTools.markdownToDrawnix': 'Markdown to Drawnix',
+
+  // Clean confirm dialog
+  'cleanConfirm.title': 'Clear Board',
+  'cleanConfirm.description':
+    'This will clear the entire board. Do you want to continue?',
+  'cleanConfirm.cancel': 'Cancel',
+  'cleanConfirm.ok': 'OK',
+
+  // Link popup items
+  'popupLink.delLink': 'Delete Link',
+
+  // Tool popup items
+  'popupToolbar.fillColor': 'Fill Color',
+  'popupToolbar.fontColor': 'Font Color',
+  'popupToolbar.link': 'Insert Link',
+  'popupToolbar.stroke': 'Stroke',
+  
+  // Text placeholders
+  'textPlaceholders.link': 'Link',
+  'textPlaceholders.text': 'Text',
+
+  // Line tool
+  'line.source': 'Start',
+  'line.target': 'End',
+  'line.arrow': 'Arrow',
+  'line.none': 'None',
+
+  // Stroke style
+  'stroke.solid': 'Solid',
+  'stroke.dashed': 'Dashed',
+  'stroke.dotted': 'Dotted',
+};
+
+export default enTranslations;
