+++ packages/drawnix/src/i18n/translations/index.ts
@@ -1,0 +1,5 @@
+import zhTranslations from './zh';
+import enTranslations from './en';
+import ruTranslations from './ru';
+
+export { zhTranslations, enTranslations, ruTranslations };
