+++ packages/drawnix/src/i18n/translations/ru.ts
@@ -1,0 +1,141 @@
+import { Translations } from '../types';
+
+const ruTranslations: Translations = {
+  // Toolbar items
+  'toolbar.hand': 'Рука — H',
+  'toolbar.selection': 'Выделение — V',
+  'toolbar.mind': 'Mind-карта — M',
+  'toolbar.text': 'Текст — T',
+  'toolbar.arrow': 'Стрелка — A',
+  'toolbar.shape': 'Фигуры',
+  'toolbar.image': 'Изображение — Cmd+U',
+  'toolbar.extraTools': 'Дополнительно',
+
+  'toolbar.pen': 'Карандаш — P',
+  'toolbar.eraser': 'Ластик — E',
+
+  'toolbar.arrow.straight': 'Прямая стрелка',
+  'toolbar.arrow.elbow': 'Ломаная стрелка',
+  'toolbar.arrow.curve': 'Кривая стрелка',
+
+  'toolbar.shape.rectangle': 'Прямоугольник — R',
+  'toolbar.shape.ellipse': 'Эллипс — O',
+  'toolbar.shape.triangle': 'Треугольник',
+  'toolbar.shape.terminal': 'Останов',
+  'toolbar.shape.diamond': 'Ромб',
+  'toolbar.shape.parallelogram': 'Параллелограмм',
+  'toolbar.shape.roundRectangle': 'Скруглённый прямоугольник',
+
+  // Zoom controls
+  'zoom.in': 'Увеличить — Cmd++',
+  'zoom.out': 'Уменьшить — Cmd+-',
+  'zoom.fit': 'По размеру экрана',
+  'zoom.100': 'Сбросить к 100%',
+  
+  // Themes
+  'theme.default': 'Стандартная',
+  'theme.colorful': 'Красочная',
+  'theme.soft': 'Мягкая',
+  'theme.retro': 'Старинная',
+  'theme.dark': 'Тёмная',
+  'theme.starry': 'Звёздная',
+
+  // Colors
+  'color.none': 'Автоматически',
+  'color.unknown': 'Другой цвет',
+  'color.default': 'Чёрный',
+  'color.white': 'Белый',
+  'color.gray': 'Серый',
+  'color.deepBlue': 'Голубой',
+  'color.red': 'Красный',
+  'color.green': 'Зелёный',
+  'color.yellow': 'Жёлтый',
+  'color.purple': 'Фиолетовый',
+  'color.orange': 'Оранжевый',
+  'color.pastelPink': 'Розовый',
+  'color.cyan': 'Лиловый',
+  'color.brown': 'Коричневый',
+  'color.forestGreen': 'Сосновный',
+  'color.lightGray': 'Светло-серый',
+
+  // General
+  'general.undo': 'Отменить',
+  'general.redo': 'Вернуть',
+  'general.menu': 'Меню приложения',
+  'general.duplicate': 'Дублировать',
+  'general.delete': 'Удалить',
+  
+  // Language
+  'language.switcher': 'Language',
+  'language.chinese': '中文',
+  'language.english': 'English',
+  'language.russian': 'Русский',
+  
+  // Menu items
+  'menu.open': 'Открыть',
+  'menu.saveFile': 'Сохранить',
+  'menu.exportImage': 'Экспортировать',
+  'menu.exportImage.png': 'PNG',
+  'menu.exportImage.jpg': 'JPG',
+  'menu.cleanBoard': 'Очистить доску',
+  'menu.github': 'GitHub',
+  
+  // Dialog translations
+  'dialog.mermaid.title': 'Mermaid в Drawnix',
+  'dialog.mermaid.description': 'Поддерживаются',
+  'dialog.mermaid.flowchart': 'блок-схемы',
+  'dialog.mermaid.sequence': 'диаграммы последовательностей', 
+  'dialog.mermaid.class': 'диаграммы классов',
+  'dialog.mermaid.otherTypes':
+    ' и другие диаграммы (преобразуются в изображения).',
+  'dialog.mermaid.syntax': 'Синтаксис Mermaid',
+  'dialog.mermaid.placeholder':
+    'Введите сюда описание вашей Mermaid-диаграммы…',
+  'dialog.mermaid.preview': 'Предпросмотр',
+  'dialog.mermaid.insert': 'Вставить',
+  'dialog.markdown.description':
+    'Поддерживается автоматическое преобразование синтаксиса Markdown в mind-карты.',
+  'dialog.markdown.syntax': 'Синтаксис Markdown',
+  'dialog.markdown.placeholder':
+    'Введите сюда описание вашего текста Markdown…',
+  'dialog.markdown.preview': 'Предпросмотр',
+  'dialog.markdown.insert': 'Вставить',
+  'dialog.error.loadMermaid': 'Не удалось загрузить библотеку Mermaid',
+  
+  // Extra tools menu items
+  'extraTools.mermaidToDrawnix': 'Mermaid в Drawnix',
+  'extraTools.markdownToDrawnix': 'Markdown в Drawnix',
+
+  // Clean confirm dialog
+  'cleanConfirm.title': 'Очистить доску',
+  'cleanConfirm.description':
+    'Это удалит всё содержимое доски. Вы хотите продолжить?',
+  'cleanConfirm.cancel': 'Отмена',
+  'cleanConfirm.ok': 'ОК',
+
+  // Link popup items
+  'popupLink.delLink': 'Удалить ссылку',
+
+  // Tool popup items
+  'popupToolbar.fillColor': 'Цвет заливки',
+  'popupToolbar.fontColor': 'Цвет текста',
+  'popupToolbar.link': 'Вставить ссылку',
+  'popupToolbar.stroke': 'Контур',
+  
+  // Text placeholders
+  'textPlaceholders.link': 'Ссылка',
+  'textPlaceholders.text': 'Текст',
+
+  // Line tool
+  'line.source': 'Начало',
+  'line.target': 'Конец',
+  'line.arrow': 'Стрелка',
+  'line.none': 'Нет',
+
+  // Stroke style
+  'stroke.solid': 'Сплошной',
+  'stroke.dashed': 'Штриховой',
+  'stroke.dotted': 'Пунктирный',
+};
+
+export default ruTranslations;
