+++ packages/drawnix/src/i18n/translations/zh.ts
@@ -1,0 +1,136 @@
+import { Translations } from '../types';
+
+const zhTranslations: Translations = {
+  // Toolbar items
+  'toolbar.hand': '手形工具 — H',
+  'toolbar.selection': '选择 — V',
+  'toolbar.mind': '思维导图 — M',
+  'toolbar.text': '文本 — T',
+  'toolbar.arrow': '箭头 — A',
+  'toolbar.shape': '形状',
+  'toolbar.image': '图片 — Cmd+U',
+  'toolbar.extraTools': '更多工具',
+
+  'toolbar.pen': '画笔 — P',
+  'toolbar.eraser': '橡皮擦 — E',
+
+  'toolbar.arrow.straight': '直线',
+  'toolbar.arrow.elbow': '肘线',
+  'toolbar.arrow.curve': '曲线',
+
+  'toolbar.shape.rectangle': '长方形 — R',
+  'toolbar.shape.ellipse': '圆 — O',
+  'toolbar.shape.triangle': '三角形',
+  'toolbar.shape.terminal': '椭圆角矩形',
+  'toolbar.shape.diamond': '菱形',
+  'toolbar.shape.parallelogram': '平行四边形',
+  'toolbar.shape.roundRectangle': '圆角矩形',
+
+  // Zoom controls
+  'zoom.in': '放大 — Cmd++',
+  'zoom.out': '缩小 — Cmd+-',
+  'zoom.fit': '自适应',
+  'zoom.100': '缩放至 100%',
+
+  // Themes
+  'theme.default': '默认',
+  'theme.colorful': '缤纷',
+  'theme.soft': '柔和',
+  'theme.retro': '复古',
+  'theme.dark': '暗夜',
+  'theme.starry': '星空',
+
+  // Colors
+  'color.none': '主题颜色',
+  'color.unknown': '其他颜色',
+  'color.default': '黑色',
+  'color.white': '白色',
+  'color.gray': '灰色',
+  'color.deepBlue': '深蓝色',
+  'color.red': '红色',
+  'color.green': '绿色',
+  'color.yellow': '黄色',
+  'color.purple': '紫色',
+  'color.orange': '橙色',
+  'color.pastelPink': '淡粉色',
+  'color.cyan': '青色',
+  'color.brown': '棕色',
+  'color.forestGreen': '森绿色',
+  'color.lightGray': '浅灰色',
+
+  // General
+  'general.undo': '撤销',
+  'general.redo': '重做',
+  'general.menu': '应用菜单',
+  'general.duplicate': '复制',
+  'general.delete': '删除',
+
+  // Language
+  'language.switcher': 'Language',
+  'language.chinese': '中文',
+  'language.english': 'English',
+  'language.russian': 'Русский',
+  
+  // Menu items
+  'menu.open': '打开',
+  'menu.saveFile': '保存文件',
+  'menu.exportImage': '导出图片',
+  'menu.exportImage.png': 'PNG',
+  'menu.exportImage.jpg': 'JPG',
+  'menu.cleanBoard': '清除画布',
+  'menu.github': 'GitHub',
+
+  // Dialog translations
+  'dialog.mermaid.title': 'Mermaid 转 Drawnix',
+  'dialog.mermaid.description': '目前仅支持',
+  'dialog.mermaid.flowchart': '流程图',
+  'dialog.mermaid.sequence': '序列图',
+  'dialog.mermaid.class': '类图',
+  'dialog.mermaid.otherTypes': '。其他类型在 Drawnix 中将以图片呈现。',
+  'dialog.mermaid.syntax': 'Mermaid 语法',
+  'dialog.mermaid.placeholder': '在此处编写 Mermaid 图表定义…',
+  'dialog.mermaid.preview': '预览',
+  'dialog.mermaid.insert': '插入',
+  'dialog.markdown.description': '支持 Markdown 语法自动转换为思维导图。',
+  'dialog.markdown.syntax': 'Markdown 语法',
+  'dialog.markdown.placeholder': '在此处编写 Markdown 文本定义…',
+  'dialog.markdown.preview': '预览',
+  'dialog.markdown.insert': '插入',
+  'dialog.error.loadMermaid': '加载 Mermaid 库失败',
+
+  // Extra tools menu items
+  'extraTools.mermaidToDrawnix': 'Mermaid 到 Drawnix',
+  'extraTools.markdownToDrawnix': 'Markdown 到 Drawnix',
+
+  // Clean confirm dialog
+  'cleanConfirm.title': '清除画布',
+  'cleanConfirm.description': '这将会清除整个画布。你是否要继续?',
+  'cleanConfirm.cancel': '取消',
+  'cleanConfirm.ok': '确认',
+
+  // Link popup items
+  'popupLink.delLink': '移除连结',
+
+  // Tool popup items
+  'popupToolbar.fillColor': '填充颜色',
+  'popupToolbar.fontColor': '字体颜色',
+  'popupToolbar.link': '链接',
+  'popupToolbar.stroke': '边框',
+  
+  // Text placeholders
+  'textPlaceholders.link': '链接',
+  'textPlaceholders.text': '文本',
+
+  // Line tool
+  'line.source': '起点',
+  'line.target': '终点',
+  'line.arrow': '箭头',
+  'line.none': '无',
+
+  // Stroke style
+  'stroke.solid': '实线',
+  'stroke.dashed': '虚线',
+  'stroke.dotted': '点线',
+};
+
+export default zhTranslations;
